Update README.md
I'm a research fellow at the University of Surrey, working on a project called SMART (Shaping Multilingual Access through Respeaking) focuses on interlingual respeaking (IRSP) as a rapidly developing practice and form of human-machine interaction that provides real-time interlingual speech-to-text transfer via speech recognition. There is a high demand for interlingual respeaking to provide access to content on television, on the internet, at live events (Moores, 2020), and in the classroom (Romero-Fresco et al., 2019) as a form of social integration and ensuring equal access to different audience groups (d/Deaf, deafened and hard of hearing and other-language speakers alike). However, little empirical research has been conducted on this complex practice to date.